LayoutTests/platform/blackberry/editing/selection/anchor-focus2-expected.txt

EDITING DELEGATE: shouldChangeSelectedDOMRange:(null) toDOMRange:range from 3 of BODY > HTML > #document to 3 of BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
This tests the anchorNode, anchorOffset, focusNode and focusOffset properties of the Selection object. These properties are part of Mozilla's Selection object API, and so their values should be consistent in both browsers.


text
Test Failed. Error was: Anchor ([object HTMLBodyElement], 3) is incorrect.
